# rest/openapi.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: REST Bookstore
  description: "Assignment #1: gRPC and REST implementation. Group: AA"
servers:
  - url: http://localhost:3000
paths:
  /books:
    get:
      summary: List all books
      operationId: list
      tags:
        - books
      responses:
        '200':
          description: Array of all books
          content:
            application/json:    
              schema:
                $ref: "#/components/schemas/Books"
    post:
      summary: Insert a book
      operationId: insert
      tags:
        - books
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
      responses:
        '201':
          description: Success response
          content:
                application/json:    
                  example: {}
  /books/{bookId}:
    get:
      summary: Info for a specific book
      operationId: get
      tags:
        - books
      parameters:
        - name: bookId
          in: path
          required: true
          description: The id of the book to retrieve
          schema:
            type: string
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        '404':
          description: Not Found
    delete:
      summary: Delete for a specific book
      operationId: delete
      tags:
        - books
      parameters:
        - name: bookId
          in: path
          required: true
          description: The id of the book to retrieve
          schema:
            type: string
      responses:
        '204':
          description: Successful delete
        '404':
          description: Not Found
components:
  schemas:
    Book:
      type: object
      required:
        - id
        - title
        - author
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        author:
          type: string
    Books:
      type: array
      items:
        $ref: "#/components/schemas/Book"